repo 3.11.2

+++ otherlibs/win32unix/channels.c
@@ -16,12 +16,13 @@
 #include <mlvalues.h>
 #include <alloc.h>
 #include <io.h>
+#include "caml_io.h"
 #include <memory.h>
 #include "unixsupport.h"
 #include <fcntl.h>
 
-extern long _get_osfhandle(int);
-extern int _open_osfhandle(long, int);
+extern intptr_t _get_osfhandle(int);
+extern int _open_osfhandle(intptr_t, int);
 
 int win_CRT_fd_of_filedescr(value handle)
 {
